remove non-wasm
