+++ README.md
@@ -1,0 +1,3 @@
+# godom
+
+Fork of [siongui's godom](https://github.com/siongui/godom) for WebAssembly only + some new methods.
