--- example/index.html
@@ -1,22 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Test Page</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-<body>
-
-<div id="info"></div>
-<input type="text" id="ip"><br>
-
-<br><br><br>
-
-<div><a class="person" href="/p1/" data-content="content of person 1">person 1</a></div>
-<div><a class="person" href="/p2/" data-content="content of person 2">person 2</a></div>
-<div><a class="person" href="/p3/" data-content="content of person 3">person 3</a></div>
-<div id="infoHistory"></div>
-
-<script src="app.js"></script>
-</body>
-</html>
